+++ csvEditorHtml/browser/beforeDomLoadedBrowser.ts
@@ -10,6 +10,9 @@ const executeAfterDomLoadedQueue: Array<() => void> = []
 
 var initialVars: InitialVars = {
 	isWatchingSourceFile: false,
+	sourceFileCursorLineIndex: null,
+	sourceFileCursorColumnIndex: null,
+	isCursorPosAfterLastColumn: false,
 }
 
 var initialConfig: EditCsvConfig | undefined = {
@@ -49,6 +52,7 @@ var initialConfig: EditCsvConfig | undefined = {
 	insertColBehavior: 'keepRowKeepColumn',
 	initiallyIsInReadonlyMode: false,
 	hideOpenCsvEditorUiActions: false, //noop, has only effect if set inside the user settings (vs code extension)
+	openTableAtCursorPos: "onlyInitially",
 }
 
 function __getById(id: string): HTMLElement {