+++ thirdParty/papaparse/papaparse.js
@@ -1,10 +1,10 @@
 /* @license
 Papa Parse
-v5.0.0
+v5.0.0-custom-1.0.1
 https://github.com/mholt/PapaParse
 License: MIT
 commit: 49170b76b382317356c2f707e2e4191430b8d495
-comment: we need this to support custom comment handling...
+fork -> https://github.com/janisdd/PapaParse/tree/fix609_main
 */
 /*
 
@@ -13,6 +13,13 @@ you need to manually compress it, e.g. with https://javascript-minifier.com/
 
 changelog: (latest first)
 
+- started to track versions with `-custom-1.0.0` suffix
+- added config options:
+  - `calcLineIndexToCsvLineIndexMapping: bool` and `calcColumnIndexToCsvColumnIndexMapping: bool`
+  - if set to true, the result will contain `outLineIndexToCsvLineIndexMapping` and `outColumnIndexToCsvColumnIndexMapping`
+    - outLineIndexToCsvLineIndexMapping: for every line in the input text the csv line it refers to
+    - outColumnIndexToCsvColumnIndexMapping: the end string indices for every every csv line fields (for every csv row)
+- fixed and issue where multi-character delimiter won't work
 - added option `quoteEmptyOrNullFields` (defaults to false) to unparse which defines how null, undefined and empty strings are quoted
 - fixes issue where all fields quoted and missing closing quote on last field will hang the function guessDelimiter
 	- this is because the field in the row will not terminated by the new line because the closing quote is missing (\n is also valid inside multi line fields)
@@ -1199,6 +1206,10 @@ changelog: (latest first)
 			_input = '';
 		};
 
+		/**
+		 * @param s {Array<string>}
+		 * @return {boolean}
+		 */
 		function testEmptyLine(s) {
 			return _config.skipEmptyLines === 'greedy' ? s.join('').trim() === '' : s.length === 1 && s[0].length === 0;
 		}
@@ -1211,11 +1222,89 @@ changelog: (latest first)
 				_delimiterError = false;
 			}
 
+			// even if skip empty lines is set, we have empty lines here, we filter them out later
+			if (_config.calcLineIndexToCsvLineIndexMapping) {
+			//true: calculate a line mapping from input text to csv lines
+				var outLineIndexToCsvLineIndexMapping = [];
+				var currentCsvLineIndex = 0;
+				var lastRealCsvLineIndex = 0;
+
+				for (var i = 0; i < _results.data.length; i++) {
+					/** @type {Array<String>} */
+					var csvLine = _results.data[i];
+
+					for (var j = 0; j < csvLine.length; j++) {
+						var csvField = csvLine[j];
+
+						//csv line cells might contain new line chars...
+						var newLinesCount = csvField.split(_config.newline).length - 1;
+
+						for (var k = 0; k < newLinesCount; k++) {
+							outLineIndexToCsvLineIndexMapping.push(currentCsvLineIndex);
+						}
+					}
+
+					outLineIndexToCsvLineIndexMapping.push(currentCsvLineIndex);
+
+					//for empty lines we want the next csv line index
+					if (_config.skipEmptyLines && testEmptyLine(csvLine)) {
+						//don't change the index
+					} else {
+						lastRealCsvLineIndex = currentCsvLineIndex;
+						currentCsvLineIndex++;
+					}
+				}
+
+				//there is a special case then the last lines are empty, and we want to skip empty lines
+				//then the csv line index of that last line should be the last csv line (before it would be an invalid index after we filter out the empty lines)
+				//e.g.
+				// before:
+				// 1,2,3  --> 0
+				//			  --> 1
+				//				--> 1
+				// 4,5,6	--> 1
+				// 7,8,9	--> 2
+				//				--> 3
+				//				--> 4
+				//---
+				// after:
+				// 1,2,3  --> 0
+				//			  --> 1
+				//				--> 1
+				// 4,5,6	--> 1
+				// 7,8,9	--> 2
+				//				--> 2 (corrected)
+				//				--> 2 (corrected)
+				if (_results.data.length > 0 && _config.skipEmptyLines) {
+
+					var correctingLineIndexIndex = outLineIndexToCsvLineIndexMapping.length - 1;
+					for (var m = _results.data.length - 1; m >= 0; m--) {
+						var _csvLine = _results.data[m];
+
+						if (testEmptyLine(_csvLine)) {
+							outLineIndexToCsvLineIndexMapping[correctingLineIndexIndex] = lastRealCsvLineIndex;
+							correctingLineIndexIndex--;
+
+						} else {
+							// after we find the first line with content, we can stop
+							break;
+						}
+					}
+				}
+
+				_results.outLineIndexToCsvLineIndexMapping = outLineIndexToCsvLineIndexMapping;
+			}
+
 			if (_config.skipEmptyLines)
 			{
-				for (var i = 0; i < _results.data.length; i++)
-					if (testEmptyLine(_results.data[i]))
-						_results.data.splice(i--, 1);
+				//see https://github.com/mholt/PapaParse/pull/912/files
+				// for (var i = 0; i < _results.data.length; i++)
+				// 	if (testEmptyLine(_results.data[i]))
+				// 		_results.data.splice(i--, 1);
+				_results.data = _results.data.filter(function(d) {
+					return !testEmptyLine(d);
+				});
+
 			}
 
 			if (needsHeaderRow())
@@ -1500,6 +1589,9 @@ changelog: (latest first)
 		//(when need to skip empty & comment rows and during this we might reset columnIsQuoted multiple times)
 		var firstQuoteInformationRowFound = false;
 
+		//true: collect the string indices for every csv field on the first row
+		var calcColumnIndexToCsvColumnIndexMapping = config.calcColumnIndexToCsvColumnIndexMapping;
+
 		// Delimiter must be valid
 		if (typeof delim !== 'string'
 			|| Papa.BAD_DELIMITERS.indexOf(delim) > -1)
@@ -1550,6 +1642,23 @@ changelog: (latest first)
 			cursor = 0;
 			var data = [], errors = [], row = [], lastCursor = 0;
 
+			//note this is the 0 based string index of the fields
+			//this also includes the separators
+			//e.g. "1,2222,33" --> [1, 6, 8]
+			//because
+			// [0,1] = "1,"
+			// [2,3,4,5,6] = "2222,"
+			// [7,8] = "33"
+			//if we us the indices we should always get the delimiter(end)
+			//can be -1 if the field is empty (because we don't skip empty lines before post-processing)(e.g. when the last line is \n)
+			//this is because the out csv line mapping includes entries for the text file lines
+			var outColumnIndexToCsvColumnIndexMapping = null;
+			var currRowColumnIndexToCsvColumnIndexMapping = [];
+
+			if (calcColumnIndexToCsvColumnIndexMapping) {
+				outColumnIndexToCsvColumnIndexMapping = [];
+			}
+
 
 			if (!input)
 				return returnable();
@@ -1586,6 +1695,11 @@ changelog: (latest first)
 							columnIsQuoted = Array(_row.length).fill(false);
 						}
 
+						//does not support calcColumnIndexToCsvColumnIndexMapping (too lazy to copy)
+						if (calcColumnIndexToCsvColumnIndexMapping) {
+							throw new Error('calcColumnIndexToCsvColumnIndexMapping not supported with stepping function');
+						}
+
 						pushRow(_row);
 						doStep();
 						if (aborted)
@@ -1597,6 +1711,25 @@ changelog: (latest first)
 						if (retainQuoteInformation && firstQuoteInformationRowFound === false) {
 							//in fast mode there are no quote characters...
 							columnIsQuoted = Array(_row.length).fill(false);
+						}
+
+						if (calcColumnIndexToCsvColumnIndexMapping) {
+							if (isCommentRow) {
+								//only one string in the row
+								currRowColumnIndexToCsvColumnIndexMapping.push(row.length - 1); //-1 to get 0 based index
+							} else {
+								//we have only delimiters...
+								var _cummulativeLength = 0;
+								// eslint-disable-next-line no-loop-func
+								_row.forEach(function(value, index) {
+									if (index !== _row.length - 1) {
+										_cummulativeLength += value.length + delimLen;
+									} else {
+										_cummulativeLength += value.length;
+									}
+									currRowColumnIndexToCsvColumnIndexMapping.push(_cummulativeLength - 1); //-1 to get 0 based index
+								});
+							}
 						}
 
 						pushRow(_row);
@@ -1617,6 +1750,9 @@ changelog: (latest first)
 			var quoteSearch = input.indexOf(quoteChar, cursor);
 			//we don't use fast mode so we assume some field is quoted...
 			columnIsQuoted = [];
+
+			var currentRowStartIndex = 0; //string index used to calculate the relative current field index in the current row
+			var currentFieldEndIndex = -1;
 
 			//if the text does not contain the delimiter (not even in quoted fields) we can return early
 			if (isGuessingDelimiter && nextDelim === -1 && cursor === 0) {
@@ -1662,6 +1798,13 @@ changelog: (latest first)
 									index: cursor
 								});
 							}
+
+							if (nextNewline === -1) {
+								addColumnIndexMapping(inputLen - 1);
+							} else {
+								addColumnIndexMapping(nextNewline - 1);
+							}
+
 							return finish();
 						}
 
@@ -1669,6 +1812,8 @@ changelog: (latest first)
 						if (quoteSearch === inputLen - 1)
 						{
 							var value = input.substring(cursor, quoteSearch).replace(quoteCharRegex, quoteChar);
+							currentFieldEndIndex = quoteSearch;
+							addColumnIndexMapping(currentFieldEndIndex);
 							return finish(value);
 						}
 
@@ -1700,6 +1845,9 @@ changelog: (latest first)
 						// Closing quote followed by delimiter or 'unnecessary spaces + delimiter'
 						if (input.substr(quoteSearch + 1 + spacesBetweenQuoteAndDelimiter, delimLen) === delim)
 						{
+							currentFieldEndIndex = quoteSearch + spacesBetweenQuoteAndDelimiter + delimLen;
+							addColumnIndexMapping(currentFieldEndIndex);
+
 							row.push(input.substring(cursor, quoteSearch).replace(quoteCharRegex, quoteChar));
 							cursor = quoteSearch + 1 + spacesBetweenQuoteAndDelimiter + delimLen;
 
@@ -1718,6 +1866,10 @@ changelog: (latest first)
 						// Closing quote followed by newline or 'unnecessary spaces + newLine'
 						if (input.substr(quoteSearch + 1 + spacesBetweenQuoteAndNewLine, newlineLen) === newline)
 						{
+							//special case for mapping because the new line is the row terminator
+							currentFieldEndIndex = quoteSearch + spacesBetweenQuoteAndNewLine;
+							addColumnIndexMapping(currentFieldEndIndex);
+
 							row.push(input.substring(cursor, quoteSearch).replace(quoteCharRegex, quoteChar));
 							saveRow(quoteSearch + 1 + spacesBetweenQuoteAndNewLine + newlineLen);
 							nextDelim = input.indexOf(delim, cursor);	// because we may have skipped the nextDelim in the quoted field
@@ -1774,10 +1926,18 @@ changelog: (latest first)
 
 					if (nextNewline === -1) {
 						//add the last comment
+
+						// eslint-disable-next-line camelcase
+						currentFieldEndIndex = input.length - 1;
+						addColumnIndexMapping(currentFieldEndIndex);
+
 						row.push(input.substring(cursor));
 						pushRow(row); // is called in finish
 						return returnable();
 					}
+
+					currentFieldEndIndex = nextNewline - 1;
+					addColumnIndexMapping(currentFieldEndIndex);
 
 					row.push(input.substring(cursor, nextNewline));
 					saveRow(nextNewline + newlineLen);
@@ -1797,6 +1957,10 @@ changelog: (latest first)
 						if (nextDelimObj && typeof nextDelimObj.nextDelim !== 'undefined') {
 							nextDelim = nextDelimObj.nextDelim;
 							quoteSearch = nextDelimObj.quoteSearch;
+
+							currentFieldEndIndex = nextDelim;
+							addColumnIndexMapping(currentFieldEndIndex);
+
 							row.push(input.substring(cursor, nextDelim));
 							cursor = nextDelim + delimLen;
 							// we look for next delimiter char
@@ -1804,6 +1968,10 @@ changelog: (latest first)
 							continue;
 						}
 					} else {
+
+						currentFieldEndIndex = nextDelim;
+						addColumnIndexMapping(currentFieldEndIndex);
+
 						row.push(input.substring(cursor, nextDelim));
 						cursor = nextDelim + delimLen;
 						nextDelim = input.indexOf(delim, cursor);
@@ -1814,6 +1982,9 @@ changelog: (latest first)
 				// End of row
 				if (nextNewline !== -1)
 				{
+					currentFieldEndIndex = nextNewline - 1;
+					addColumnIndexMapping(currentFieldEndIndex);
+
 					row.push(input.substring(cursor, nextNewline));
 					saveRow(nextNewline + newlineLen);
 
@@ -1836,6 +2007,8 @@ changelog: (latest first)
 				break;
 			}
 
+			currentFieldEndIndex = input.length - 1;
+			addColumnIndexMapping(currentFieldEndIndex);
 
 			return finish();
 
@@ -1848,6 +2021,12 @@ changelog: (latest first)
 				data.push(row);
 				lastCursor = cursor;
 
+				if (calcColumnIndexToCsvColumnIndexMapping) {
+					outColumnIndexToCsvColumnIndexMapping.push(currRowColumnIndexToCsvColumnIndexMapping);
+					currRowColumnIndexToCsvColumnIndexMapping = [];
+				}
+				currentRowStartIndex = cursor;
+
 				if (firstQuoteInformationRowFound === false) {
 
 					if (row.length === 1 &&
@@ -1858,6 +2037,16 @@ changelog: (latest first)
 					} else {
 						firstQuoteInformationRowFound = true;
 					}
+				}
+			}
+
+			/**
+			 * adds the given index to the column mapping if we still calculate it
+			 * @param cumulativeColumnIndex
+			 */
+			function addColumnIndexMapping(cumulativeColumnIndex) {
+				if (calcColumnIndexToCsvColumnIndexMapping) {
+					currRowColumnIndexToCsvColumnIndexMapping.push(cumulativeColumnIndex - currentRowStartIndex);
 				}
 			}
 
@@ -1912,7 +2101,8 @@ changelog: (latest first)
 			function returnable(stopped, step)
 			{
 				var isStep = step || false;
-				return {
+
+				var result = {
 					data: isStep ? data[0]  : data,
 					errors: errors,
 					meta: {
@@ -1924,6 +2114,13 @@ changelog: (latest first)
 					},
 					columnIsQuoted: columnIsQuoted
 				};
+
+				// use config because we use calcColumnIndexToCsvColumnIndexMapping to notify top
+				if (config.calcColumnIndexToCsvColumnIndexMapping) {
+					result.outColumnIndexToCsvColumnIndexMapping = outColumnIndexToCsvColumnIndexMapping
+				}
+
+				return result;
 			}
 
 			/** Executes the user's step function and resets data & errors. */